refactor: simplify decode logic
